server/requirements.txt
antlr4-python3-runtime==4.10
anyio==3.6.1
asgiref==3.5.2
astor==0.8.1
attrs==21.4.0
authlib
certifi==2022.6.15
charset-normalizer
click==8.1.3
dnspython==2.2.1
docker
email-validator==1.2.1
fastapi==0.78.0
fastapi_utils
fastapi-sessions==0.3.2
graphviz==0.20
h11==0.13.0
httpx
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
lxml
packaging==21.3
pluggy==1.0.0
psycopg2
py==1.11.0
pydantic
pyparsing==3.0.9
pytest==7.1.2
python-cas==1.6.0
requests~=2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
staticfg==0.9.5
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
websocket-client==1.3.3
pydot
fastapi_utils
